requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-decouple==3.8
sqlparse==0.5.3
django-storages==1.14.6
boto3==1.39.17
Pillow==11.3.0
pillow-heif
